requirements.txt
azure-kusto-ingest==4.0.1
